updating product/profile img

+++ routes/productsRoutes.js
@@ -28,9 +28,13 @@ router.get('/', async (req, res) => {
 router.get('/:id', isAuth, async (req, res) => {
   try {
     const product = await Product.findById(req.params.id)
+    let img = product.img
+    if (img != '/images/empty.png') {
+      img = '/' + img
+    }
     const userInSession = await User.findById(req.user._id)
 
-    res.render('product', { product, userInSession })
+    res.render('product', { product, userInSession, img })
   } catch (e) {
     console.log(e)
   }
@@ -66,8 +70,10 @@ router.post('/edit', isAuth, async (req, res) => {
     delete req.body.id //щоб передати все оновлене окрім id-його залишити
 
     if (req.file) {
-      var filePath = `.${req.body.img}` //видаляєм старе фото з бази
-      fs.unlinkSync(filePath)
+      if (req.body.img != '/images/empty.png') {
+        var filePath = `./${req.body.img}` //видаляєм старе фото з бази
+        fs.unlinkSync(filePath)
+      }
       req.body.img = req.file.path
     }
 
@@ -88,6 +94,21 @@ router.post('/edit', isAuth, async (req, res) => {
 
 router.post('/:id/remove', isAuth, async (req, res) => {
   const products = await Product.deleteOne({ _id: req.body.productId })
+
+  if (req.body.img != '/images/empty.png') {
+    var filePath = `./${req.body.img}` //видаляєм старе фото з бази
+    fs.unlinkSync(filePath)
+  }
+
+  //видалення продукту і з модельки користувача
+  const user = await User.findById(req.body.userId)
+  let i = 0
+  user.products.forEach((element) => {
+    if (element == req.body.productId) user.products.splice(i, 1)
+    ++i
+  })
+  await user.save()
+
   alert.type = 'success'
   alert.message = 'Your product has been successfully deleted'
   res.redirect('/products')
